+++ src/main/java/com/imtoocai/diary/service/ChangePassword.java
@@ -3,7 +3,7 @@ package com.imtoocai.diary.service;
 import com.imtoocai.diary.model.Result;
 
 public interface ChangePassword {
-    Result ChangePassword(String email, Integer oldPassword, Integer newPassword);
+    Result ChangePassword(String email, Long oldPassword, Long  newPassword);
 
-    Result ChangePassword(Integer userId, Integer oldPassword, Integer newPassword);
+    Result ChangePassword(Integer userId, Long oldPassword, Long newPassword);
 }
