+++ src/main/java/com/imtoocai/diary/service/ResetPassword.java
@@ -3,6 +3,6 @@ package com.imtoocai.diary.service;
 import com.imtoocai.diary.model.Result;
 
 public interface ResetPassword {
-    Result resetPassword(String email, String code, Integer newPassword);
+    Result resetPassword(String email, String code, Long newPassword);
 
 }
